[#1] Setup Next Card Show/Hide Feature
