+++ src/App.js
@@ -1,29 +1,7 @@
 import React from "react";
 import users from "./data/user";
 import Card from "./components/Card";
-
-const getItems = (acc, obj, items, id) => {
-  if (obj.manager === id) {
-    obj.children = getTree(items, obj.id);
-    acc.push(obj);
-  }
-  return acc;
-};
-
-const getTree = (items, manager) => {
-  const mapper = [];
-  for (let i = 0, t = items.length; i < t; i++) {
-    const item = items[i];
-    if (item.manager === manager) {
-      item.children = items.reduce(
-        (acc, obj) => getItems(acc, obj, items, item.id),
-        []
-      );
-      mapper.push(item);
-    }
-  }
-  return mapper;
-};
+import { getTree } from "./utils/parsers";
 
 function App() {
   return (
